backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
openai>=1.14.0
pinecone>=3.2.2
pydantic>=2.7.4
python-multipart>=0.0.6
gunicorn>=21.2.0